K8SPSMDB-651: Recover from full cluster crash
If all pods of replica set removed when the cluster is exposed with
LoadBalancer, they fail the join to the replica set again. The operator
sees the replica set as initialized but fail to connect to it due to
non-existent primary.

To fix the issue, if operator sees the replica set as initialized but
couldn't connect to it, it increases the priority of the first pod until
replica set selects it as primary. Afterwards the operator reverts the
priority to the default value.